+++ scripts/const.js
@@ -117,6 +117,10 @@ export const SOURCES = {
     MANIFEST_EIDOLON: "Compendium.pf2e.actionspf2e.Item.n5vwBnLSlIXL9ptp",
   },
 
+  THAUMATURGE: {
+    MIRRORS_REFLECTION: "Compendium.pf2e.actionspf2e.Item.Mh4Vdg6gu8g8RAjh",
+  },
+
   //SF2e
   MECHANIC: {
     DEPLOY_MINE:
@@ -150,6 +154,7 @@ export const SLUG_TO_SOURCE = {
   light: SOURCES.MISC.LIGHT,
   "living-graveyard": SOURCES.NECROMANCER.LIVING_GRAVEYARD,
   "manifest-eidolon": SOURCES.SUMMONER.MANIFEST_EIDOLON,
+  "mirrors-reflection": SOURCES.THAUMATURGE.MIRRORS_REFLECTION,
   "perfected-thrall": SOURCES.NECROMANCER.PERFECTED_THRALL,
   "phantasmal-minion": SOURCES.MISC.PHANTASMAL_MINION,
   "plant-banner": SOURCES.COMMANDER.PLANT_BANNER,
@@ -396,6 +401,10 @@ export const EFFECTS = {
       img: "icons/magic/death/grave-tombstone-glow-teal.webp",
     }),
   },
+  THAUMATURGE: {
+    MIRRORS_REFLECTION:
+      "Compendium.pf2e-summons-assistant.pf2e-summons-assistant-items.Item.1I6uoL93lNAUZQ1t",
+  },
   COMMANDER: {
     IN_PLANT_BANNER_RANGE:
       "Compendium.pf2e-summons-assistant.pf2e-summons-assistant-items.Item.vnFV2b3aYdvGeVkM",
@@ -559,6 +568,10 @@ export const ACTIONS = {
     GUARD:
       "Compendium.pf2e-summons-assistant.pf2e-summons-assistant-items.Item.JdCfKdOAdumgw6aU",
     PUSH: "Compendium.pf2e-summons-assistant.pf2e-summons-assistant-items.Item.cyjLIVsxZbJNr4PF",
+  },
+  THAUMATURGE: {
+    SHATTER_REFLECTION:
+      "Compendium.pf2e-summons-assistant.pf2e-summons-assistant-items.Item.mv56YnVvWLylqJWM",
   },
 };