+++ scripts/settings.js
@@ -118,6 +118,16 @@ export function setupSettings() {
     type: Boolean,
   });
 
+  game.settings.register(MODULE_ID, "specific-case.handle.thaumaturge", {
+    name: `${MODULE_ID}.module-settings.specific-case.handle.thaumaturge.name`,
+    hint: `${MODULE_ID}.module-settings.specific-case.handle.thaumaturge.hint`,
+    requiresReload: true,
+    scope: "world",
+    config: true,
+    default: true,
+    type: Boolean,
+  });
+
   game.settings.register(MODULE_ID, "disabled-items", {
     name: `${MODULE_ID}.module-settings.disabled-items.name`,
     hint: `${MODULE_ID}.module-settings.disabled-items.hint`,