+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 ## Unreleased
 
+## 2.6.0
+
+- **New**
+  - `Mirror Implement` Thaumaturge support
+- **Updated**
+  - Added a default pick for `Dancing Blade`
+  - Added a way in specific summons to exclude default summoned traits from some summon
+
 ## 2.5.4
 
 - Fixed bug introduced in `2.2.0` that was causing summons such as `Wall of Stone` to break due to issue with function `crosshairParameters` (🐛 @mgoldstein322)